feat(context): context support a setter
Using data = {state: ‘blah’} it is now possible to set the context of
the current application. I use a typical query interface here where a
method can be used as a getter or a setter, provided that:
- callback is the last arguments
- if two arguments are given, the method is used as a setter.

We made some research in my last company and it appeared that
javascript dev prefer to have an API with less methods with multiple
signature and are very accustomed to jQuery paradigm.